+++ e2e/docs/.vuepress/layouts/Category.vue
@@ -38,7 +38,7 @@ const categoryMap = useBlogCategory('category')
 @use '../styles/mixins.scss';
 
 .category-wrapper {
-  @include mixins.content_wrapper;
+  @include mixins.content-wrapper;
 
   padding-top: 1rem !important;
   padding-bottom: 0 !important;