+++ themes/theme-default/src/client/components/global/Badge.vue
@@ -1,30 +1,27 @@
 <script setup lang="ts">
-defineProps({
-  type: {
-    type: String,
-    required: false,
-    default: 'tip',
+import type { VNode } from 'vue'
+
+withDefaults(
+  defineProps<{
+    type?: string
+    text?: string
+    vertical?: string
+  }>(),
+  {
+    type: 'tip',
+    text: '',
+    vertical: undefined,
   },
-  text: {
-    type: String,
-    required: false,
-    default: '',
-  },
-  vertical: {
-    type: String,
-    required: false,
-    default: undefined,
-  },
-})
+)
 
 defineSlots<{
-  default?: () => any
+  default?: () => VNode | VNode | string | null
 }>()
 </script>
 
 <template>
   <span
-    class="badge"
+    class="vp-badge"
     :class="type"
     :style="{
       verticalAlign: vertical,
@@ -33,3 +30,48 @@ defineSlots<{
     <slot>{{ text }}</slot>
   </span>
 </template>
+
+<style lang="scss">
+.vp-badge {
+  display: inline-block;
+  vertical-align: top;
+
+  height: 18px;
+  padding: 0 6px;
+  border-radius: 3px;
+
+  color: var(--c-bg);
+
+  font-weight: 600;
+  font-size: 14px;
+  line-height: 18px;
+
+  transition:
+    color var(--t-color),
+    background-color var(--t-color);
+
+  &.tip {
+    background-color: var(--c-badge-tip);
+  }
+
+  &.warning {
+    background-color: var(--c-badge-warning);
+    color: var(--c-badge-warning-text);
+  }
+
+  &.danger {
+    background-color: var(--c-badge-danger);
+    color: var(--c-badge-danger-text);
+  }
+
+  // update the vertical align for the badge in toc
+  .table-of-contents & {
+    vertical-align: middle;
+  }
+
+  // avoid extra space between badges
+  & + & {
+    margin-left: 5px;
+  }
+}
+</style>