+++ themes/theme-default/src/client/components/global/CodeGroupItem.vue
@@ -3,25 +3,14 @@ defineOptions({
   name: 'CodeGroupItem',
 })
 
-defineProps({
-  title: {
-    type: String,
-    required: true,
-  },
-  active: {
-    type: Boolean,
-    required: false,
-    default: false,
-  },
-})
+defineProps<{
+  title: string
+  active?: boolean
+}>()
 </script>
 
 <template>
-  <div
-    class="code-group-item"
-    :class="{ 'code-group-item__active': active }"
-    role="tabpanel"
-  >
+  <div class="code-group-item" :class="{ active }" role="tabpanel">
     <slot />
   </div>
 </template>