+++ themes/theme-default/src/client/composables/useNavigate.ts
@@ -1,8 +1,6 @@
 import { isLinkAbsolute, isLinkWithProtocol } from '@vuepress/helper/client'
 import { useRoute, useRouter } from 'vuepress/client'
 
-const FAKE_HOST = 'http://.'
-
 export const useNavigate = (): ((url: string) => void) => {
   const router = useRouter()
   const route = useRoute()
@@ -13,13 +11,11 @@ export const useNavigate = (): ((url: string) => void) => {
         // Inner absolute path
         if (route.path !== url) router.push(url)
       } else if (isLinkWithProtocol(url)) {
-        // Outer url
-        if (window) window.open(url)
+        // external url
+        window?.open(url)
       } else {
-        // Inner relative path
-        const loc = route.path.slice(0, route.path.lastIndexOf('/'))
-
-        router.push(new URL(`${loc}/${encodeURI(url)}`, FAKE_HOST).pathname)
+        // relative url
+        router.push(encodeURI(url))
       }
   }
 }