+++ themes/theme-default/src/client/config.ts
@@ -1,13 +1,11 @@
+import { setupDarkMode } from '@theme/useDarkMode'
+import { useScrollPromise } from '@theme/useScrollPromise'
+import { setupSidebarItems } from '@theme/useSidebarItems'
 import { hasGlobalComponent } from '@vuepress/helper/client'
 import { h } from 'vue'
 import { defineClientConfig } from 'vuepress/client'
 import type { ClientConfig } from 'vuepress/client'
 import { Badge, CodeGroup, CodeGroupItem } from './components/global/index.js'
-import {
-  setupDarkMode,
-  setupSidebarItems,
-  useScrollPromise,
-} from './composables/index.js'
 import Layout from './layouts/Layout.vue'
 import NotFound from './layouts/NotFound.vue'
 
@@ -22,7 +20,7 @@ export default defineClientConfig({
       app.component('CodeGroupItem', CodeGroupItem)
 
     // compat with @vuepress/plugin-docsearch and @vuepress/plugin-search
-    app.component('NavbarSearch', () => {
+    app.component('VPSearch', () => {
       const SearchComponent =
         app.component('Docsearch') || app.component('SearchBox')
       if (SearchComponent) {