+++ tools/highlighter-helper/src/node/index.ts
@@ -1,0 +1,1 @@
+export * from './lineNumbers/index.js'